Implement LiquidatedOtherItemMapper
* Mapper transforms certificate for liquidated company into table
